// frontend/src/api/model/eventStatus.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * CandiDash API
 * Job application tracking system API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Status of scheduled event.
 */
export type EventStatus = typeof EventStatus[keyof typeof EventStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EventStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  rescheduled: 'rescheduled',
  cancelled: 'cancelled',
  completed: 'completed',
} as const;