// frontend/src/api/model/documentFormat.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * CandiDash API
 * Job application tracking system API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Document format enumeration.
 */
export type DocumentFormat = typeof DocumentFormat[keyof typeof DocumentFormat];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DocumentFormat = {
  pdf: 'pdf',
  doc: 'doc',
  docx: 'docx',
  rtf: 'rtf',
  txt: 'txt',
  md: 'md',
  ppt: 'ppt',
  pptx: 'pptx',
  odp: 'odp',
  xls: 'xls',
  xlsx: 'xlsx',
  ods: 'ods',
  csv: 'csv',
  tsv: 'tsv',
  odt: 'odt',
  jpg: 'jpg',
  jpeg: 'jpeg',
  png: 'png',
  gif: 'gif',
  webp: 'webp',
  json: 'json',
  external: 'external',
} as const;