// frontend/src/api/model/applicationType.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * CandiDash API
 * Job application tracking system API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Type of application.
 */
export type ApplicationType = typeof ApplicationType[keyof typeof ApplicationType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ApplicationType = {
  job_posting: 'job_posting',
  spontaneous: 'spontaneous',
  reached_out: 'reached_out',
} as const;