// frontend/src/api/model/applicationStatus.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * CandiDash API
 * Job application tracking system API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Application status enumeration.
 */
export type ApplicationStatus = typeof ApplicationStatus[keyof typeof ApplicationStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ApplicationStatus = {
  pending: 'pending',
  follow_up_scheduled: 'follow_up_scheduled',
  interview_scheduled: 'interview_scheduled',
  rejected: 'rejected',
  accepted: 'accepted',
  obsolete: 'obsolete',
} as const;